fix map switch bug

+++ time_tasks.py
@@ -11,7 +11,7 @@ import os
 import schedule
 import json
 from datetime import datetime
-from shutil import copyfile
+import shutil
 import threading
 import schedule
 from watchdog.observers import Observer
@@ -34,13 +34,13 @@ TASK_LIST = list()
 
 
 class FileChangeHandler(FileSystemEventHandler):
-    global FILESTATUS_LOCK,TASKFILE_STATUS,MAPS_DB_PATH
     def on_modified(self, event):
+        global FILESTATUS_LOCK,TASKFILE_STATUS,MAPS_DB_PATH
         with FILESTATUS_LOCK:
             taskfile_path = MAPS_DB_PATH+"/timeTask.json"
             if event.src_path == taskfile_path:
                 TASKFILE_STATUS = 1
-                print(f'event type: {event.event_type}  path : {event.src_path}')
+                print('event type: {event.event_type}  path : {event.src_path}')
 
 def status_update_cb(status):
     global NAV_STATUS
@@ -49,17 +49,27 @@ def status_update_cb(status):
 
 def change_map(map_name, path_name):
     global MAPS_DB_PATH, CURRENT_DB_PATH
-
+    rospy.set_param("/system_monitor/nav_is_enabled", False)
     #先拷贝地图
     map_src_path =  MAPS_DB_PATH + "/" + map_name
 
     if map_name == "":
+        rospy.set_param("/system_monitor/nav_is_enabled", True)
         return False
 
     if path_name == "":
+        rospy.set_param("/system_monitor/nav_is_enabled", True)
         return False
 
     if not os.path.exists(map_src_path):
+        rospy.set_param("/system_monitor/nav_is_enabled", True)
+        return False
+
+    #如果nav_check 或者 path_check 失败，则不切换
+    nav_check = os.path.exists(map_src_path + "/nav_" + path_name + ".csv")
+    path_check = os.path.exists(map_src_path + "/path_" + path_name + ".csv")
+    if not nav_check or not path_check:
+        rospy.set_param("/system_monitor/nav_is_enabled", True)
         return False
 
     if os.path.exists(CURRENT_DB_PATH):
@@ -73,12 +83,17 @@ def change_map(map_name, path_name):
 
     nav_dst = CURRENT_DB_PATH + "/nav.csv"
     new_nav_dst = CURRENT_DB_PATH + "/new_nav.csv"
-    path_dst = = CURRENT_DB_PATH + "/path.csv"
+    path_dst  = CURRENT_DB_PATH + "/path.csv"
 
-    shutil.copy(nav_src, nav_dst)
-    shutil.copy(new_nav_src, new_nav_dst)
-    shutil.copy(path_nav_src, path_nav_dst)
 
+    if os.path.exists(nav_src):
+        shutil.copy(nav_src, nav_dst)
+    if os.path.exists(new_nav_src):
+        shutil.copy(new_nav_src, new_nav_dst)
+    if os.path.exists(path_src):
+        shutil.copy(path_src, path_dst)
+
+    rospy.set_param("/system_monitor/nav_is_enabled", True)
     return True
 
 
@@ -100,60 +115,68 @@ class MapSwitchTask():
     def ifneed_runnow(self):
         time_now = datetime.now()
         time_now_value = time_now.hour*60 + time_now.minute
-
-        if self.whichDay == 0 or self.whichDay == time_now.day:
+        #print "now "+ str(time_now_value) + " task " + str(self.startTime_value)
+        #print "daynow "+ str(time_now.day) + " daytask " + str(self.whichDay)
+        if self.whichDay == 0 or self.whichDay == (time_now.weekday() + 1):
             if time_now_value >= self.startTime_value and time_now_value<self.stopTime_value:
                 return True
 
         return False
 
     def job_needdone(self):
-        global NAV_STATUS,AUDIO_PUB，GALILEO_PUB
-            AUDIO_PUB.publish("开始切换地图")
-            #切换地图过程，发布导航服务开启禁用话题
-            STATUS_LOCK.acquire()
-            nav_status = NAV_STATUS
-            STATUS_LOCK.release()
+        global NAV_STATUS,AUDIO_PUB,GALILEO_PUB
+        AUDIO_PUB.publish("开始切换地图")
+        #切换地图过程，发布导航服务开启禁用话题
+        STATUS_LOCK.acquire()
+        nav_status = NAV_STATUS
+        STATUS_LOCK.release()
 
-            if change_map(self.mapName, self.pathName):
-                if nav_status != 1:
-                    AUDIO_PUB.publish("地图切换成功")
-                    return
-
-                #todo 地图切换完成后要把正在进行的导航任务进行重置
-                #先关闭
-                AUDIO_PUB.publish("地图切换成功, 开始重新载入导航任务")
-                galileo_cmds.data = 'm' + chr(0x04)
-                galileo_cmds.length = len(galileo_cmds.data)
-                galileo_cmds.header.stamp = rospy.Time.now()
-                GALILEO_PUB.publish(galileo_cmds)
-                max_do = 0
-                while nav_status == 1:
-                    time.sleep(3)
-                    STATUS_LOCK.acquire()
-                    nav_status = NAV_STATUS
-                    STATUS_LOCK.release()
-                    max_do = max_do +1;
-                    if max_do >20:
-                        break
-                #再重新开启
-                nav_status = 0
-                galileo_cmds.data = 'm' + chr(0x00)
-                galileo_cmds.length = len(galileo_cmds.data)
-                galileo_cmds.header.stamp = rospy.Time.now()
-                GALILEO_PUB.publish(galileo_cmds)
-                max_do = 0
-                while nav_status != 1:
-                    time.sleep(3)
-                    STATUS_LOCK.acquire()
-                    nav_status = NAV_STATUS
-                    STATUS_LOCK.release()
-                    max_do = max_do +1;
-                    if max_do >20:
-                        break
+        if change_map(self.mapName, self.pathName):
+            if nav_status != 1:
+                AUDIO_PUB.publish("地图切换成功")
                 return
-            else:
-                AUDIO_PUB.publish("地图切换失败")
+
+            #todo 地图切换完成后要把正在进行的导航任务进行重置
+            #先关闭
+            rospy.set_param("/system_monitor/nav_is_enabled", False)
+            AUDIO_PUB.publish("地图切换成功, 开始重新载入导航任务")
+            galileo_cmds = GalileoNativeCmds()
+            galileo_cmds.data = 'm' + chr(0x04)
+            galileo_cmds.length = len(galileo_cmds.data)
+            galileo_cmds.header.stamp = rospy.Time.now()
+            GALILEO_PUB.publish(galileo_cmds)
+            max_do = 0
+            while nav_status == 1:
+                time.sleep(3)
+                STATUS_LOCK.acquire()
+                nav_status = NAV_STATUS
+                STATUS_LOCK.release()
+                max_do = max_do +1;
+                if max_do >20:
+                    break
+            #再重新开启
+            rospy.set_param("/system_monitor/nav_is_enabled", True)
+            nav_status = 0
+            galileo_cmds.data = 'm' + chr(0x00)
+            galileo_cmds.length = len(galileo_cmds.data)
+            galileo_cmds.header.stamp = rospy.Time.now()
+            GALILEO_PUB.publish(galileo_cmds)
+            max_do = 0
+            while nav_status != 1:
+                time.sleep(3)
+                STATUS_LOCK.acquire()
+                nav_status = NAV_STATUS
+                STATUS_LOCK.release()
+                max_do = max_do +1;
+                if max_do >20:
+                    break
+            return
+        else:
+            AUDIO_PUB.publish("地图切换失败")
+
+def job_test(test_str):
+    global AUDIO_PUB
+    AUDIO_PUB.publish(test_str)
 
 
 def load_tasks():
@@ -171,31 +194,36 @@ def load_tasks():
             TASK_LIST.append(MapSwitchTask(task["whichDay"],task["startTime"], task["stopTime"], task["mapName"], task["pathName"]))
 
             if task["whichDay"] == 0:
-                schedule.every().day.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 1:
-                schedule.every().monday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 2:
-                schedule.every().tuesday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 3:
-                schedule.every().wednesday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 4:
-                schedule.every().thursday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 5:
-                schedule.every().friday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 6:
-                schedule.every().saturday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
-            elif task.whichDay == 7:
-                schedule.every().sunday.at(task["startTime"]+":00").do(TASK_LIST[index_task].job_needdone()).tag("mapSwithTask",str(index_task))
+                #schedule.every().day.at(task["startTime"]).do(job_test,"测试")
+                schedule.every().day.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 1:
+                schedule.every().monday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 2:
+                schedule.every().tuesday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 3:
+                schedule.every().wednesday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 4:
+                schedule.every().thursday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 5:
+                schedule.every().friday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 6:
+                schedule.every().saturday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
+            elif task["whichDay"] == 7:
+                schedule.every().sunday.at(task["startTime"]).do(TASK_LIST[index_task].job_needdone)
             index_task = index_task + 1
 
         index_task = 0
         for job_now in schedule.jobs:
             if TASK_LIST[index_task].ifneed_runnow() :
                 #任务需要立即运行
+                #print "runnow " + str(index_task)
                 job_now.run()
+            index_task = index_task + 1
 
 if __name__ == "__main__":
-    rospy.init_node("status_audio_reporter")
+
+    rospy.init_node("time_task_server")
+
     AUDIO_PUB = rospy.Publisher("/xiaoqiang_tts/text", String, queue_size=10)
 
     GALILEO_PUB = rospy.Publisher("/galileo/cmds", GalileoNativeCmds, queue_size=5)
@@ -203,6 +231,7 @@ if __name__ == "__main__":
     status_sub = rospy.Subscriber("/galileo/status", GalileoStatus, status_update_cb)
 
     #开始从json文件中加载地图切换任务
+    time.sleep(1)
     load_tasks()
     index_i = 0
 
@@ -215,7 +244,7 @@ if __name__ == "__main__":
         #任务最多每分钟执行一次
         schedule.run_pending()
         #每10分钟检查一下任务是否要重新载入
-        if index_i >10:
+        if index_i >=3:
             index_i = 0
             FILESTATUS_LOCK.acquire()
             if TASKFILE_STATUS == 1:
